+++ react/QuickOrder.tsx
@@ -1,30 +1,31 @@
 /* eslint-disable no-console */
 import React, { useState, useContext } from 'react'
-import { Button, ToastContext } from 'vtex.styleguide'
+import { Button, ToastContext, Dropdown } from 'vtex.styleguide'
 import {
   FormattedMessage,
-  injectIntl,
   defineMessages,
   WrappedComponentProps,
 } from 'react-intl'
 import { usePixel } from 'vtex.pixel-manager/PixelContext'
-import { useMutation } from 'react-apollo'
+import { useMutation, graphql, compose } from 'react-apollo'
 import { OrderForm } from 'vtex.order-manager'
 import { usePWA } from 'vtex.store-resources/PWAContext'
 import { addToCart as ADD_TO_CART } from 'vtex.checkout-resources/Mutations'
-import { compareObjects } from './modules/compareObjects'
 import { useCssHandles } from 'vtex.css-handles'
+
 import TranslatedTitle from './components/TranslatedTitle'
 import AutocompleteBlock from './components/AutocompleteBlock'
 import TextAreaBlock from './components/TextAreaBlock'
 import UploadBlock from './components/UploadBlock'
 import CategoryBlock from './components/CategoryBlock'
 import ReviewBlock from './components/ReviewBlock'
+import getSellers from './queries/sellers.gql'
 import styles from './styles.css'
 import { GetText } from './utils'
 
 const QuickOrder: StorefrontFunctionComponent<QuickOrderProps &
-  WrappedComponentProps> = ({
+  WrappedComponentProps &
+  unknown> = ({
   customToastUrl,
   title,
   showCopyPaste,
@@ -40,6 +41,7 @@ const QuickOrder: StorefrontFunctionComponent<QuickOrderProps &
   uploadText,
   uploadDescription,
   downloadText,
+  data,
   intl,
 }: any) => {
   const { showToast } = useContext(ToastContext)
@@ -50,6 +52,14 @@ const QuickOrder: StorefrontFunctionComponent<QuickOrderProps &
     textAreaValue: '',
     reviewItems: [],
     refidLoading: null,
+    seller: window.sessionStorage?.getItem('sellerId') ?? null,
+    sellers:
+      data.sellers?.itemsReturned?.map((sellerItem: any) => {
+        return {
+          value: sellerItem.id,
+          label: sellerItem.name,
+        }
+      }) || [],
   })
 
   const {
@@ -58,13 +68,15 @@ const QuickOrder: StorefrontFunctionComponent<QuickOrderProps &
     textAreaValue,
     reviewItems,
     refidLoading,
+    seller,
+    sellers,
   } = state
 
   const { push } = usePixel()
   const { settings = {}, showInstallPrompt = undefined } = usePWA() || {}
   const { promptOnCustomEvent } = settings
 
-  const { orderForm, setOrderForm }: OrderFormContext = OrderForm.useOrderForm()
+  const { setOrderForm }: OrderFormContext = OrderForm.useOrderForm()
 
   const messages = defineMessages({
     success: {
@@ -119,24 +131,29 @@ const QuickOrder: StorefrontFunctionComponent<QuickOrderProps &
     { error: mutationError, loading: mutationLoading },
   ] = useMutation<{ addToCart: OrderForm }, { items: [] }>(ADD_TO_CART)
 
+  const setSeller = (value: string) => {
+    setState({
+      ...state,
+      seller: value,
+    })
+    window.sessionStorage.setItem('sellerId', value)
+  }
+
   const callAddToCart = async (items: any) => {
     const mutationResult = await addToCart({
-      variables: { items },
+      variables: {
+        items: items.map((item: any) => {
+          return {
+            ...item,
+            seller: seller || sellers[0].value,
+          }
+        }),
+      },
     })
 
     if (mutationError) {
       console.error(mutationError)
       toastMessage({ success: false, isNewItem: false })
-      return
-    }
-    if (
-      mutationResult.data &&
-      compareObjects(mutationResult.data.addToCart, orderForm)
-    ) {
-      toastMessage({ success: true, isNewItem: false })
-      setTimeout(() => {
-        window.location.href = '/checkout'
-      }, 1000)
       return
     }
 
@@ -156,7 +173,23 @@ const QuickOrder: StorefrontFunctionComponent<QuickOrderProps &
       items: pixelEventItems,
     })
 
-    toastMessage({ success: true, isNewItem: true })
+    if (
+      mutationResult.data?.addToCart?.messages?.generalMessages &&
+      mutationResult.data.addToCart.messages.generalMessages.length
+    ) {
+      mutationResult.data.addToCart.messages.generalMessages.map((msg: any) => {
+        return showToast({
+          message: msg.text,
+          action: undefined,
+          duration: 30000,
+        })
+      })
+    } else {
+      toastMessage({ success: true, isNewItem: true })
+      setTimeout(() => {
+        window.location.href = '/checkout'
+      }, 1000)
+    }
 
     if (promptOnCustomEvent === 'addToCart' && showInstallPrompt) {
       showInstallPrompt()
@@ -170,7 +203,6 @@ const QuickOrder: StorefrontFunctionComponent<QuickOrderProps &
       return {
         id: parseInt(vtexSku, 10),
         quantity: parseFloat(quantity),
-        seller: '1',
       }
     })
     callAddToCart(items)
@@ -212,6 +244,7 @@ const QuickOrder: StorefrontFunctionComponent<QuickOrderProps &
     'reviewBlock',
     'categoryBlock',
     'buttonsBlock',
+    'sellerContainer',
   ] as const
   const handles = useCssHandles(CSS_HANDLES)
 
@@ -219,8 +252,31 @@ const QuickOrder: StorefrontFunctionComponent<QuickOrderProps &
     <div
       className={`${styles.container} ${handles.container} flex flex-column pv6 ph4`}
     >
-      <div className={`title-container ${handles.title}`}>
-        <TranslatedTitle title={title} />
+      <div
+        className="c-on-base flex flex-wrap flex-row justify-between
+            mt0"
+      >
+        <div
+          className={`vtex-pageHeader__title t-heading-2 order-0 flex-grow-1 title-container ${handles.title}`}
+        >
+          <TranslatedTitle title={title} />
+        </div>
+        <div className="vtex-pageHeader__children order-2 order-0-ns mt0-ns">
+          {sellers.length > 1 && (
+            <div className={`mt5 ${handles.sellerContainer}`}>
+              <FormattedMessage id="store/quickorder.seller" />:
+              <Dropdown
+                variation="inline"
+                size="large"
+                options={sellers}
+                value={seller}
+                onChange={(_, v) => {
+                  setSeller(v)
+                }}
+              />
+            </div>
+          )}
+        </div>
       </div>
 
       {!reviewState && showCopyPaste && (
@@ -430,4 +486,4 @@ QuickOrder.schema = {
   },
 }
 
-export default injectIntl(QuickOrder)
+export default compose(graphql(getSellers))(QuickOrder)