+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Feedback message from the checkout
+- Seller selector if the store have more than one
+
+### Fixed
+
+- Seller always being set as `1`
+
 ## [0.8.2] - 2020-06-10
 
 ### Added